Manejo de Insomnia para crear, listar, borrar y consultar por codigo

+++ tutoria-services/src/main/java/com/devsenior/jquiguantar/tutoria_services/services/StudentService.java
@@ -17,6 +17,14 @@ public class StudentService {
     public List<StudentDto> getAll() {
         return repository.findAll();
     }
+
+    public StudentDto getById(Long code){
+        if(code < 1){
+            throw new RuntimeException("El codigo enviado no es valido");
+        }
+        return repository.findByCode(code);
+    }
+
     public StudentDto save(StudentDto student) {
         if(student.getCode() == null){
             throw new RuntimeException("El campo Codigo es obligatorio: ");
@@ -27,9 +35,15 @@ public class StudentService {
         {
             throw new RuntimeException("El campo Genero es obligatorio:");
         }
+        if(student.getActive() == null){
+            student.setActive(true);
+        }
         return repository.save(student);
     }
     public StudentDto delete(Long code) {
+        if(code < 1){
+            throw new RuntimeException("El codigo enviado no es valido");
+        }
         return repository.delete(code);
     }
 }
